generic ButtJoint, Format, Lint, Test, CL
